backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
drf-yasg==1.21.7
django-filter==23.5 